# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is a user-friendly solution designed to serve as the foundation for more complex APIs.",
        "title": "Go API",
        "contact": {
            "name": "Raul del Aguila",
            "email": "email@email.com"
        }
    },
    "basePath": "/",
    "paths": {
...
        "/user": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get users",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "profile_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "updated_at",
                        "example": "'updated_at', 'created_at', 'name' or some other field from response object",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.ItemsOutputDTO-github_com_raulaguila_go-api_internal_pkg_dto_UserOutputDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Insert user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Insert user",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "User model",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.UserInputDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.UserOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete user by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "User ID",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.IDsInputDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/pass": {
            "put": {
                "description": "Set user password by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Set user password",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "User email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Password model",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.PasswordInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Reset user password by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Reset user password",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "User email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get user by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.UserOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update user by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User model",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.UserInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.UserOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}/photo": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get user photo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user photo",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Set user photo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Set user photo",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "profile photo",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_pkg_helper.HTTPResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_raulaguila_go-api_internal_pkg_dto.IDsInputDTO": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_pkg_dto.ItemsOutputDTO-github_com_raulaguila_go-api_internal_pkg_dto_UserOutputDTO": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.UserOutputDTO"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.PaginationDTO"
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_pkg_dto.UserInputDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.cena@email.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Cena"
                },
                "profile_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_raulaguila_go-api_internal_pkg_dto.UserOutputDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.cena@email.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "John Cena"
                },
                "profile": {
                    "$ref": "#/definitions/github_com_raulaguila_go-api_internal_pkg_dto.ProfileOutputDTO"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_raulaguila_go-api_pkg_helper.HTTPResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}